pyproject.toml
[project]
name = "pydantic-ui"
version = "0.1.0"
description = "A textual editor and validator for YAML/JSON files against Pydantic models."
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pydantic",
    "pyyaml",
]

[project.scripts]
pydantic-ui = "pydantic_ui.__main__:main"
